+++ src/main/kotlin/lotto/domin/LottoNumberSet.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domin
 
 data class LottoNumberSet(
     val lotto: List<Int>
