+++ src/main/kotlin/lotto/dto/WinningStaticsResponseDto.kt
@@ -1,4 +1,6 @@
-package lotto
+package lotto.dto
+
+import lotto.domin.LottoWinningAmount
 
 data class WinningStaticsResponseDto(
     val winningLottoList: Map<LottoWinningAmount, Int>,
