+++ src/main/kotlin/lotto/util/LottoNumberGenerator.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.util
 
 interface LottoNumberGenerator {
